# libs/contracts-api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: MikasMissions Public API
  version: 1.0.0
  description: Canonical v1 API contract for parent and child product flows.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/kids/home:
    get:
      operationId: getKidsHome
      tags: [Kids]
      parameters:
        - $ref: '#/components/parameters/ChildProfileIDQuery'
        - name: mode
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/KidsMode'
      responses:
        '200':
          description: Kids home payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KidsHomeResponse'
        '400':
          $ref: '#/components/responses/APIError'
  /v1/kids/progress/{child_profile_id}:
    get:
      operationId: getKidsProgress
      tags: [Kids]
      parameters:
        - $ref: '#/components/parameters/ChildProfileIDPath'
      responses:
        '200':
          description: Kids progress payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KidsProgressResponse'
        '400':
          $ref: '#/components/responses/APIError'
...
  /v1/parents/controls/{child_profile_id}:
    get:
      operationId: getParentControls
      tags: [Parents]
      parameters:
        - $ref: '#/components/parameters/ChildProfileIDPath'
      responses:
        '200':
          description: Current parental controls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentControlsResponse'
        '400':
          $ref: '#/components/responses/APIError'
    put:
      operationId: putParentControls
      tags: [Parents]
      parameters:
        - $ref: '#/components/parameters/ChildProfileIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentalControls'
      responses:
        '200':
          description: Updated parental controls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentControlsResponse'
        '400':
          $ref: '#/components/responses/APIError'
...
  /v1/admin/workflows/{workflow_id}:
    put:
      operationId: updateAdminWorkflow
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/WorkflowIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminWorkflowRequest'
      responses:
        '200':
          description: Workflow template updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWorkflow'
        '400':
          $ref: '#/components/responses/APIError'
        '404':
          $ref: '#/components/responses/APIError'
    delete:
      operationId: deleteAdminWorkflow
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/WorkflowIDPath'
      responses:
        '204':
          description: Workflow template deleted.
        '400':
          $ref: '#/components/responses/APIError'
        '404':
          $ref: '#/components/responses/APIError'
...
  /v1/admin/model-profiles/{id}:
    get:
      operationId: getAdminModelProfile
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/ModelProfileIDPath'
      responses:
        '200':
          description: Model profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminModelProfile'
        '400':
          $ref: '#/components/responses/APIError'
        '404':
          $ref: '#/components/responses/APIError'
    put:
      operationId: putAdminModelProfile
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/ModelProfileIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminModelProfile'
      responses:
        '200':
          description: Model profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminModelProfile'
        '400':
          $ref: '#/components/responses/APIError'
components:
  parameters:
    ChildProfileIDPath:
      name: child_profile_id
      in: path
      required: true
      schema:
        type: string
    ChildProfileIDQuery:
      name: child_profile_id
      in: query
      required: true
      schema:
        type: string
    WorkflowIDPath:
      name: workflow_id
      in: path
      required: true
      schema:
        type: string
...
    ModelProfileIDPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    APIError:
      description: API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    AgeBand:
      type: string
      enum: [3-5, 6-11, 12-16]
    SafetyMode:
      type: string
      enum: [strict, balanced]
    KidsMode:
      type: string
      enum: [early, core, teen]
...
    RailItem:
      type: object
      required:
        - episode_id
        - title
        - summary
        - thumbnail_url
        - duration_ms
        - age_band
        - content_suitability
        - learning_tags
        - reason_code
        - safety_applied
        - age_fit_score
      properties:
        episode_id:
          type: string
        title:
          type: string
        summary:
          type: string
        thumbnail_url:
          type: string
          format: uri
        duration_ms:
          type: integer
        age_band:
          $ref: '#/components/schemas/AgeBand'
        content_suitability:
          type: string
          enum: [early, core, teen]
        learning_tags:
          type: array
          items:
            type: string
        reason_code:
          type: string
        safety_applied:
          type: boolean
        age_fit_score:
          type: number
...
    KidsHomeResponse:
      type: object
      required: [child_profile_id, mode, safety_mode, primary_actions, rails]
      properties:
        child_profile_id:
          type: string
        mode:
          $ref: '#/components/schemas/KidsMode'
        safety_mode:
          $ref: '#/components/schemas/SafetyMode'
        primary_actions:
          type: array
          items:
            type: string
        rails:
          type: array
          items:
            $ref: '#/components/schemas/RailItem'
...
    KidsProgressResponse:
      type: object
      required:
        - child_profile_id
        - watched_minutes_today
        - watched_minutes_7d
        - completion_percent
        - mission_streak_days
        - session_limit_minutes
        - session_minutes_used
        - session_capped
        - last_episode_id
      properties:
        child_profile_id:
          type: string
        watched_minutes_today:
          type: integer
        watched_minutes_7d:
          type: integer
        completion_percent:
          type: integer
        mission_streak_days:
          type: integer
        session_limit_minutes:
          type: integer
        session_minutes_used:
          type: integer
        session_capped:
          type: boolean
        last_episode_id:
          type: string
...
    ParentalControls:
      type: object
      required: [autoplay, chat_enabled, external_links, session_limit_minutes, bedtime_window, safety_mode]
      properties:
        autoplay:
          type: boolean
        chat_enabled:
          type: boolean
        external_links:
          type: boolean
        session_limit_minutes:
          type: integer
        bedtime_window:
          type: string
        safety_mode:
          $ref: '#/components/schemas/SafetyMode'
    ParentControlsResponse:
      type: object
      required: [child_profile_id, controls]
      properties:
        child_profile_id:
          type: string
        controls:
          $ref: '#/components/schemas/ParentalControls'
...
    AdminWorkflow:
      type: object
      required:
        - workflow_id
        - name
        - description
        - content_suitability
        - age_band
        - steps
        - model_profile_id
        - safety_profile
        - version
      properties:
        workflow_id:
          type: string
        name:
          type: string
        description:
          type: string
        content_suitability:
          type: string
          enum: [early, core, teen]
        age_band:
          $ref: '#/components/schemas/AgeBand'
        steps:
          type: array
          items:
            type: string
        model_profile_id:
          type: string
        safety_profile:
          type: string
        version:
          type: integer
          minimum: 1
...
    CreateAdminWorkflowRequest:
      type: object
      required: [name, description, content_suitability, age_band, steps, model_profile_id, safety_profile]
      properties:
        name:
          type: string
        description:
          type: string
        content_suitability:
          type: string
          enum: [early, core, teen]
        age_band:
          $ref: '#/components/schemas/AgeBand'
        steps:
          type: array
          items:
            type: string
        model_profile_id:
          type: string
        safety_profile:
          type: string
    UpdateAdminWorkflowRequest:
      allOf:
        - $ref: '#/components/schemas/CreateAdminWorkflowRequest'
...
    AdminModelProfile:
      type: object
      required: [model_profile_id, provider, base_url, model_id, timeout_ms, max_retries, safety_preset]
      properties:
        model_profile_id:
          type: string
        provider:
          type: string
          enum: [nvidia_nim]
        base_url:
          type: string
          format: uri
        model_id:
          type: string
        timeout_ms:
          type: integer
        max_retries:
          type: integer
        safety_preset:
          type: string